requirements.txt
fastapi
uvicorn
openai
pydantic
python-dotenv
pytest
slowapi 
python-multipart